tools/requirements_local_pipeline.txt
# Local run: CV DefectFill + optional SDXLRefiner (annotate / test_pipeline)
# Python 3.10+ recommended. Install PyTorch with CUDA from https://pytorch.org/ first if using --sdxl.

numpy<2.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0

# SDXL refiner (skip if you only run without --sdxl)
torch>=2.1.0
torchvision>=0.16.0
diffusers==0.29.2
transformers==4.38.0
accelerate==0.27.0
huggingface-hub==0.23.4
safetensors
